+++ packages/pyroscope-flamegraph/src/convert/sandwichViewProfiles.ts
@@ -1,18 +1,5 @@
-/* eslint-disable import/prefer-default-export */
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
-
 import type { Flamebearer } from '@pyroscope/models/src';
-import { flamebearersToTree } from './flamebearersToTree';
-
-export interface TreeNode {
-  name: string;
-  key: string;
-  self: [number];
-  total: [number];
-  offset?: number;
-  children: TreeNode[];
-}
+import { flamebearersToTree, TreeNode } from './flamebearersToTree';
 
 interface FlamebearerData {
   maxSelf: number;
@@ -51,38 +38,40 @@ export const treeToFlamebearer = (tree: TreeNode): FlamebearerData => {
   return flamebearerData;
 };
 
-const arrayToTree = (
-  nodesArray: TreeNode[],
-  maxLvlNumber: number,
-  total: number
-): TreeNode => {
-  const result = {};
+const arrayToTree = (nodesArray: TreeNode[], total: number): TreeNode => {
+  const result = {} as TreeNode;
   let nestedObj = result;
-  const emptyLvls = maxLvlNumber - nodesArray.length;
-
-  for (let i = 0; i < emptyLvls; i += 1) {
-    const emptyNode = {
-      name: '',
-      key: '',
-      total: [0],
-      self: [0],
-      children: [],
-      offset: total,
-    };
-
-    nestedObj.children = [emptyNode];
-    nestedObj = emptyNode;
-  }
 
   nodesArray.forEach(({ name, ...rest }) => {
     const nextNode = { name, ...rest, total: [total] };
-
     nestedObj.children = [nextNode];
     nestedObj = nextNode;
   });
 
   return result.children[0];
 };
+
+function dedupTree(node: TreeNode) {
+  const childrenMap: { [key: string]: TreeNode } = {};
+  for (let i = 0; i < node.children.length; i += 1) {
+    childrenMap[node.children[i].name] ||= node.children[i];
+  }
+  for (let i = 0; i < node.children.length; i += 1) {
+    const currentNode = node.children[i];
+    const existingNode = childrenMap[node.children[i].name];
+    if (existingNode !== currentNode) {
+      existingNode.total[0] += currentNode.total[0];
+      existingNode.self[0] += currentNode.self[0];
+      existingNode.children = existingNode.children.concat(
+        currentNode.children
+      );
+    }
+  }
+  node.children = Object.values(childrenMap);
+  for (let i = 0; i < node.children.length; i += 1) {
+    dedupTree(node.children[i]);
+  }
+}
 
 export function calleesFlamebearer(
   f: Flamebearer,
@@ -106,8 +95,8 @@ export function calleesFlamebearer(
     total: [],
     self: [0],
     children: [],
-  };
-  const processTree = (node) => {
+  } as TreeNode;
+  const processTree = (node: TreeNode) => {
     if (node.name === nodeName) {
       result.numTicks += node.total[0];
 
@@ -119,6 +108,7 @@ export function calleesFlamebearer(
     }
   };
   processTree(tree);
+  dedupTree(totalNode);
 
   return { ...result, ...treeToFlamebearer(totalNode) };
 }
@@ -139,9 +129,8 @@ export function callersFlamebearer(
     spyName: f.spyName,
   };
 
-  const targetFunctionTotals = [];
-  const subtrees = [];
-  let maxSubtreeLvl = 0;
+  const targetFunctionTotals: number[] = [];
+  const subtrees: TreeNode[][] = [];
 
   const totalNode = {
     name: nodeName,
@@ -149,18 +138,16 @@ export function callersFlamebearer(
     total: [0],
     self: [0],
     children: [],
-  };
-  const processTree = (node, parentNodes = []) => {
-    const currentSubtree = parentNodes.concat([{ ...node, children: [] }]);
+  } as TreeNode;
+  const processTree = (node: TreeNode, parentNodes: TreeNode[] = []) => {
+    const currentSubtree = parentNodes.concat([
+      { ...node, children: [] } as TreeNode,
+    ]);
 
     if (node.name === nodeName) {
       subtrees.push(currentSubtree);
       targetFunctionTotals.push(node.total[0]);
       result.numTicks += node.total[0];
-
-      if (maxSubtreeLvl < currentSubtree.length) {
-        maxSubtreeLvl = currentSubtree.length;
-      }
     }
 
     for (let i = 0; i < node.children.length; i += 1) {
@@ -169,25 +156,27 @@ export function callersFlamebearer(
   };
   processTree(tree);
 
+  // 1. we first make a regular tree
   subtrees.forEach((v, i) => {
-    totalNode.children.push(
-      arrayToTree(v, maxSubtreeLvl, targetFunctionTotals[i])
-    );
+    totalNode.children.push(arrayToTree(v.reverse(), targetFunctionTotals[i]));
   });
 
-  return { ...result, ...treeToFlamebearer(totalNode) };
+  // 2. that allows us to use the same dedup function
+  dedupTree(totalNode);
+
+  const flamebearer = treeToFlamebearer(totalNode);
+
+  // 3. then we reverse levels so that tree goes from bottom to top
+  flamebearer.levels = flamebearer.levels.reverse().slice(0, -1);
+
+  return { ...result, ...flamebearer };
 }
 
 export function calleesProfile(f: Flamebearer, nodeName: string): Flamebearer {
   const copy = JSON.parse(JSON.stringify(f));
   const calleesResultFlamebearer = calleesFlamebearer(copy, nodeName);
 
-  return {
-    version: 1,
-    // flamebearer: calleesResultFlamebearer,
-    // metadata: copy.metadata,
-    ...calleesResultFlamebearer,
-  };
+  return calleesResultFlamebearer;
 }
 
 export function callersProfile(f: Flamebearer, nodeName: string): Flamebearer {
@@ -195,10 +184,5 @@ export function callersProfile(f: Flamebearer, nodeName: string): Flamebearer {
 
   const callersResultFlamebearer = callersFlamebearer(copy, nodeName);
 
-  return {
-    version: 1,
-    // flamebearer: callersResultFlamebearer,
-    // metadata: copy.metadata,
-    ...callersResultFlamebearer,
-  };
+  return callersResultFlamebearer;
 }