+++ pkg/storage/types.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/dgraph-io/badger/v2"
 	"github.com/pyroscope-io/pyroscope/pkg/storage/cache"
+	"github.com/pyroscope-io/pyroscope/pkg/storage/metadata"
 	"github.com/pyroscope-io/pyroscope/pkg/util/bytesize"
 )
 
@@ -130,4 +131,14 @@ type BadgerDBWithCache interface {
 	DBInstance() *badger.DB
 	CacheInstance() *cache.Cache
 	Name() string
+}
+
+type ApplicationMetadata struct {
+	// Fully Qualified Name. Eg app.cpu ({__name__}.{profile_type})
+	FQName string `gorm:"index,unique;not null;default:null" json:"name"`
+
+	SpyName         string                   `json:"spyName,omitempty"`
+	SampleRate      uint32                   `json:"sampleRate,omitempty"`
+	Units           metadata.Units           `json:"units,omitempty"`
+	AggregationType metadata.AggregationType `json:"-"`
 }