+++ webapp/javascript/services/tags.ts
@@ -6,12 +6,28 @@ import {
 } from '@webapp/models/tags';
 import { request, parseResponse } from './base';
 
-export async function fetchTags(query: string) {
-  const response = await request(`/labels?query=${query}`);
+export async function fetchTags(query: string, from: number, until: number) {
+  const params = new URLSearchParams({
+    query,
+    from: from.toString(10),
+    until: until.toString(10),
+  });
+  const response = await request(`/labels?${params.toString()}`);
   return parseResponse<Tags>(response, TagsSchema);
 }
 
-export async function fetchLabelValues(label: string, query: string) {
-  const response = await request(`/label-values?label=${label}&query=${query}`);
+export async function fetchLabelValues(
+  label: string,
+  query: string,
+  from: number,
+  until: number
+) {
+  const params = new URLSearchParams({
+    query,
+    label,
+    from: from.toString(10),
+    until: until.toString(10),
+  });
+  const response = await request(`/label-values?${params.toString()}`);
   return parseResponse<TagsValues>(response, TagsValuesSchema);
 }